git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@1666 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/db/exporter/XlsRowDataConverter.java
@@ -88,7 +88,7 @@ public class XlsRowDataConverter
 		{
 			HSSFSheet info = wb.createSheet("SQL");
 			HSSFRow infoRow = info.createRow(0);
-			HSSFCell cell = infoRow.createCell(0);
+			HSSFCell cell = infoRow.createCell((short)0);
 			
 			HSSFCellStyle style = wb.createCellStyle();
 			style.setAlignment(HSSFCellStyle.ALIGN_LEFT);
